general clean up of '_fourier' vs not, updated a few images
