+++ final/scripts/normality_final.py
@@ -48,7 +48,7 @@ from Image_Visualizing import make_mask, present_3d
 
 # Progress bar
 toolbar_width=len(sub_list)
-sys.stdout.write("Shaprio-Wilk test for normality:  ")
+sys.stdout.write("Normality, :  ")
 sys.stdout.write("[%s]" % (" " * toolbar_width))
 sys.stdout.flush()
 sys.stdout.write("\b" * (toolbar_width+1)) # return to start of line, after '['
@@ -58,7 +58,7 @@ unmasked_prop = [] # Unmasked (all voxels)
 masked_prop = [] # Masked. 
 
 for i in sub_list:
-    residuals =   np.load(residual_data+i+"_residual.npy")
+    residuals =   np.load(residual_data+i+"_residual_fourier.npy")
     sw_pvals = check_sw(residuals)
     unmasked_prop.append(np.mean(sw_pvals > 0.05))
 
@@ -68,17 +68,6 @@ for i in sub_list:
     masked_pvals = make_mask(sw_pvals, mask_data, fit=True)
     pvals_in_brain = sw_pvals.ravel()[masked_pvals.ravel() != 0]
     masked_prop.append(np.mean(pvals_in_brain > 0.05))
-    
-    if (i[-3:]=="010"): 
-        # Save image plots of unmasked p-values for subject 10. 
-        plt.imshow(present_3d(sw_pvals), cmap=plt.get_cmap('gray'))
-        plt.savefig(location_of_images+i+'sw.png')
-        plt.close()
-
-        # Save image plots of masked p-values for a single subject. 
-        plt.imshow(present_3d(masked_pvals), cmap=plt.get_cmap('gray'))
-        plt.savefig(location_of_images+i+'swmasked.png')
-        plt.close()
      
     sys.stdout.write("-")
     sys.stdout.flush()
@@ -91,6 +80,16 @@ plt.close()
 plt.hist(np.array(masked_prop))
 plt.title("Histogram of Proportions for each Subject")
 plt.savefig(location_of_images+'maskedhist.png')
+plt.close()
+
+# Save image plots of unmasked p-values for a single subject. 
+plt.imshow(present_3d(sw_pvals), cmap=plt.get_cmap('gray'))
+plt.savefig(location_of_images+i+'sw.png')
+plt.close()
+
+# Save image plots of masked p-values for a single subject. 
+plt.imshow(present_3d(masked_pvals), cmap=plt.get_cmap('gray'))
+plt.savefig(location_of_images+i+'swmasked.png')
 plt.close()
